feat: Update docparser script to accept input directory
This commit updates the `docparser.py` script to accept an input directory as an option. This allows users to specify the directory containing the documents to convert. The file paths are then obtained from the input directory using the `glob` function. This change improves the flexibility and usability of the script.