Better batching, handling of default model, simple flow to get responses (#6)
* feat: Add a simple flow to get granite model responses

Signed-off-by: shiv-sr <shiv.sr@ibm.com>

* refactor: Split dataset into smaller batches for improved processing

This commit modifies the  class in  to split the dataset into smaller batches for improved processing. The code now uses start and end ids of the batch to split dataset and creates a batch on the fly rather than creating the batches apriori. This is necessary for large datasets.

Signed-off-by: shiv-sr <shiv.sr@ibm.com>

* refactor: Update LLMBlock to use default model if model_id is not provided

This commit modifies the LLMBlock class in llmblock.py to use the default model if the model_id is not provided. It checks if the model_id is None and if so, retrieves the default model id from the client. This ensures that the LLMBlock can still function even if the model_id is not explicitly specified.

Signed-off-by: shiv-sr <shiv.sr@ibm.com>

* refactor: Remove commented out code for dataset batching

The code changes in sdg.py remove the commented out code that was used for dataset batching. This code is no longer needed and can be safely removed.

Signed-off-by: shiv-sr <shiv.sr@ibm.com>

---------

Signed-off-by: shiv-sr <shiv.sr@ibm.com>