fix: Update llmblock.py to use self.model instead of model_id
This commit modifies the llmblock.py file to update the LLMBlock class. It changes the code to use the `self.model` attribute instead of the `model_id` parameter when calling the `server_supports_batched` function. This ensures that the correct model is used for checking if the server supports batched requests.

Signed-off-by: shiv-sr <shiv.sr@ibm.com>