github: shallow-clone for actionlint and markdown check jobs
For GitHub Actions that invoke setuptools-scm, we need the full Git
history in order to get the proper versioning values from "git
describe".

For GitHub Actions that do not invoke setuptools-scm, we do not need the
full Git history, so we can speed up those clone operations with
shallow clones.